pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.2.0",
    "langchain>=0.2.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.98",
    "pinecone>=5.0.0",
    "pydantic>=2.0.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
    "google-genai>=0.8.0",
]
